+++ js/examples/openai/backend/src/routes/feedback.route.ts
@@ -1,0 +1,8 @@
+import express, { Router } from "express";
+import { feedback } from "../controllers/feedback.controller";
+
+const router: Router = express.Router();
+
+router.route("/").post(feedback);
+
+export default router;